test: nightly live-probe CI to catch advertised-but-stubbed tools
The default CI suite is deterministic and offline, so it cannot detect a tool
that claims live behaviour but returns stub data — exactly how check_watches
shipped a fake price of 0 undetected. This adds the missing automation layer:

- .github/workflows/live-probes.yml: runs `TRVL_TEST_LIVE_PROBES=1 go test
  -run Probe ./...` nightly (+ workflow_dispatch). A re-stubbed checker returns
  0 and fails; a rotated provider API (e.g. Wizz version path) also fails here.
- Reorganise the new probes into convention `_probe_test.go` files with "Probe"
  in the name so `-run Probe` discovers them:
  - internal/livecheck/livecheck_probe_test.go (TestChecker_LiveFlightProbe)
  - mcp/tools_watch_price_probe_test.go (TestHandleCheckWatches_LiveProbe)
  Deterministic end-to-end tests stay in the offline suite.
- Broaden `make test-live-probes` from internal/flights+internal/hotels to
  `./...` so livecheck + mcp probes are included locally too.

Verified: deterministic suite green (probes self-skip); `-run Probe` discovers
both probes; both PASS live (real HEL-LHR price end-to-end).
