+++ .github/workflows/live-probes.yml
@@ -1,0 +1,31 @@
+name: Live Probes
+
+# Nightly (and on-demand) execution of the live provider probe tests. The default
+# CI suite is deterministic and offline, which means it cannot catch a tool that
+# is advertised as live but silently returns stub data (the class of bug that let
+# check_watches report a fake price of 0 for weeks). These probes hit the real
+# provider endpoints, so a re-stubbed checker or a rotated provider API (e.g. the
+# Wizz Air version path) fails here and surfaces as a tracked failure.
+
+on:
+  schedule:
+    # 05:17 UTC daily (off the hour to avoid Actions congestion).
+    - cron: "17 5 * * *"
+  workflow_dispatch: {}
+
+permissions:
+  contents: read
+
+jobs:
+  live-probes:
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+      - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6.4.0
+        with:
+          go-version: "1.26.4"
+      - name: Live provider probes
+        # -run Probe selects every *_probe_test.go test across the module; the
+        # env gate flips them from skipped to live.
+        run: TRVL_TEST_LIVE_PROBES=1 go test -count=1 -run Probe ./...
